+++ src/app/api/calendar/route.ts
@@ -6,11 +6,27 @@ export async function GET(req: NextRequest) {
     const { data, error } = await supabase.from("calendar").select("*");
 
     if (error) {
-      NextResponse.json(error);
+      NextResponse.json({ error: error.message });
     }
-    console.log(data);
     return NextResponse.json(data);
   } catch (error) {
     console.log("supabase error", error);
+  }
+}
+
+export async function POST(req: NextRequest) {
+  try {
+    const values: any = await req.json();
+    console.log(values);
+
+    const { data, error } = await supabase.from("test_calendar").insert(values);
+
+    if (error) {
+      NextResponse.json({ error: error.message });
+    }
+    return NextResponse.json(data);
+  } catch (error: any) {
+    console.log("post error", error);
+    throw new Error(error.message);
   }
 }
