+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { AppProps } from "next/app";
 import { useAuthStore } from "@/store/auth";
 import { createClient } from "@/utils/supabase/server";
-import "@/styles/globals.css";
+import "@/app/globals.css";
 
 function MyApp({ Component, pageProps }: AppProps) {
   const setUser = useAuthStore((state) => state.setUser);
