+++ src/utils/supabase/server.ts
@@ -1,29 +1,29 @@
-import { createServerClient, type CookieOptions } from "@supabase/ssr";
-import { cookies } from "next/headers";
+// import { createServerClient, type CookieOptions } from "@supabase/ssr";
+// import { cookies } from "next/headers";
 
-export function createClient() {
-  const cookieStore = cookies();
+// export function createClient() {
+//   const cookieStore = cookies();
 
-  return createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        getAll() {
-          return cookieStore.getAll();
-        },
-        setAll(cookiesToSet) {
-          try {
-            cookiesToSet.forEach(({ name, value, options }) =>
-              cookieStore.set(name, value, options)
-            );
-          } catch {
-            // The `setAll` method was called from a Server Component.
-            // This can be ignored if you have middleware refreshing
-            // user sessions.
-          }
-        },
-      },
-    }
-  );
-}
+//   return createServerClient(
+//     process.env.NEXT_PUBLIC_SUPABASE_URL!,
+//     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+//     {
+//       cookies: {
+//         getAll() {
+//           return cookieStore.getAll();
+//         },
+//         setAll(cookiesToSet) {
+//           try {
+//             cookiesToSet.forEach(({ name, value, options }) =>
+//               cookieStore.set(name, value, options)
+//             );
+//           } catch {
+//             // The `setAll` method was called from a Server Component.
+//             // This can be ignored if you have middleware refreshing
+//             // user sessions.
+//           }
+//         },
+//       },
+//     }
+//   );
+// }
